refactor(models): convert remaining barrel imports
