+++ src/models/SetBonus.ts
@@ -1,4 +1,4 @@
-import { Client } from '@/client'
+import { Client } from '@/client/Client'
 import { Artifact } from '@/models/Artifact'
 
 /**
